+++ examples/ns-metadata/resources.py
@@ -1,0 +1,31 @@
+from Pyro4 import expose
+
+
+class Resource(object):
+    pass
+
+
+@expose()
+class LaserPrinter(Resource):
+    pass
+
+
+@expose()
+class MatrixPrinter(Resource):
+    pass
+
+
+@expose()
+class PhotoPrinter(Resource):
+    pass
+
+
+@expose()
+class TapeStorage(Resource):
+    pass
+
+
+@expose()
+class DiskStorage(Resource):
+    pass
+