+++ src/Pyro4/naming_storage.py
@@ -45,12 +45,37 @@ class SqlStorage(MutableMapping):
             raise ValueError("We don't support the sqlite :memory: database type. Just use the default volatile in-memory store.")
         self.dbfile = dbfile
         with closing(sqlite3.connect(dbfile)) as db:
-            db.execute("""CREATE TABLE IF NOT EXISTS pyro_names
-                (
-                    name nvarchar PRIMARY KEY NOT NULL,
-                    uri nvarchar NOT NULL
-                );""")
+            db.execute("PRAGMA foreign_keys=ON")
+            try:
+                db.execute("SELECT COUNT(*) FROM pyro_names").fetchone()
+            except sqlite3.OperationalError:
+                # the table does not yet exist
+                self._create_schema(db)
+            else:
+                # check if we need to update the existing schema
+                try:
+                    db.execute("SELECT COUNT(*) FROM pyro_metadata").fetchone()
+                except sqlite3.OperationalError:
+                    # metadata schema needs to be created and existing data migrated
+                    db.execute("ALTER TABLE pyro_names RENAME TO pyro_names_old")
+                    self._create_schema(db)
+                    db.execute("INSERT INTO pyro_names(name, uri) SELECT name, uri FROM pyro_names_old")
+                    db.execute("DROP TABLE pyro_names_old")
             db.commit()
+
+    def _create_schema(self, db):
+        db.execute("""CREATE TABLE pyro_names
+            (
+                id integer PRIMARY KEY,
+                name nvarchar NOT NULL UNIQUE,
+                uri nvarchar NOT NULL
+            );""")
+        db.execute("""CREATE TABLE pyro_metadata
+            (
+                object integer NOT NULL,
+                metadata nvarchar NOT NULL,
+                FOREIGN KEY(object) REFERENCES pyro_names(id)
+            );""")
 
     def __getattr__(self, item):
         raise NotImplementedError("SqlStorage doesn't implement method/attribute '"+item+"'")
@@ -58,19 +83,33 @@ class SqlStorage(MutableMapping):
     def __getitem__(self, item):
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
-                result = db.execute("SELECT uri FROM pyro_names WHERE name=?", (item,)).fetchone()
+                result = db.execute("SELECT id, uri FROM pyro_names WHERE name=?", (item,)).fetchone()
                 if result:
-                    return result[0]
+                    dbid, uri = result
+                    metadata = {m[0] for m in db.execute("SELECT metadata FROM pyro_metadata WHERE object=?", (dbid,)).fetchall()}
+                    return uri, metadata
                 else:
                     raise KeyError(item)
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in getitem: "+str(e))
 
     def __setitem__(self, key, value):
+        uri, metadata = value
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
-                db.execute("DELETE FROM pyro_names WHERE name=?", (key,))
-                db.execute("INSERT INTO pyro_names(name, uri) VALUES(?,?)", (key, value))
+                cursor = db.cursor()
+                cursor.execute("PRAGMA foreign_keys=ON")
+                dbid = cursor.execute("SELECT id FROM pyro_names WHERE name=?", (key,)).fetchone()
+                if dbid:
+                    dbid = dbid[0]
+                    cursor.execute("DELETE FROM pyro_metadata WHERE object=?", (dbid,))
+                    cursor.execute("DELETE FROM pyro_names WHERE id=?", (dbid,))
+                cursor.execute("INSERT INTO pyro_names(name, uri) VALUES(?,?)", (key, uri))
+                if metadata:
+                    object_id = cursor.lastrowid
+                    for m in metadata:
+                        cursor.execute("INSERT INTO pyro_metadata(object, metadata) VALUES (?,?)", (object_id, m))
+                cursor.close()
                 db.commit()
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in setitem: "+str(e))
@@ -92,7 +131,12 @@ class SqlStorage(MutableMapping):
     def __delitem__(self, key):
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
-                db.execute("DELETE FROM pyro_names WHERE name=?", (key,))
+                db.execute("PRAGMA foreign_keys=ON")
+                dbid = db.execute("SELECT id FROM pyro_names WHERE name=?", (key,)).fetchone()
+                if dbid:
+                    dbid = dbid[0]
+                    db.execute("DELETE FROM pyro_metadata WHERE object=?", (dbid,))
+                    db.execute("DELETE FROM pyro_names WHERE id=?", (dbid,))
                 db.commit()
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in delitem: "+str(e))
@@ -108,43 +152,77 @@ class SqlStorage(MutableMapping):
     def clear(self):
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
+                db.execute("PRAGMA foreign_keys=ON")
+                db.execute("DELETE FROM pyro_metadata")
                 db.execute("DELETE FROM pyro_names")
                 db.execute("VACUUM")
                 db.commit()
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in clear: "+str(e))
 
-    def optimized_prefix_list(self, prefix):
+    def optimized_prefix_list(self, prefix, return_metadata=False):
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
-                result = db.execute("SELECT name, uri FROM pyro_names WHERE name LIKE ?", (prefix+'%',))
                 names = {}
-                for name, uri in result.fetchall():
-                    names[name] = uri
+                if return_metadata:
+                    for dbid, name, uri in db.execute("SELECT id, name, uri FROM pyro_names WHERE name LIKE ?", (prefix+'%',)).fetchall():
+                        metadata = {m[0] for m in db.execute("SELECT metadata FROM pyro_metadata WHERE object=?", (dbid,)).fetchall()}
+                        names[name] = uri, metadata
+                else:
+                    for name, uri in db.execute("SELECT name, uri FROM pyro_names WHERE name LIKE ?", (prefix+'%',)).fetchall():
+                        names[name] = uri
                 return names
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in optimized_prefix_list: "+str(e))
 
-    def optimized_regex_list(self, regex):
+    def optimized_regex_list(self, regex, return_metadata=False):
         # defining a regex function isn't much better than simply regexing ourselves over the full table.
         return None
 
+    def optimized_metadata_search(self, metadata, return_metadata=False):
+        try:
+            with closing(sqlite3.connect(self.dbfile)) as db:
+                params = list(metadata)
+                params.append(len(metadata))
+                sql = "SELECT id, name, uri FROM pyro_names WHERE id IN (SELECT object FROM pyro_metadata WHERE metadata IN ({seq}) " \
+                      "GROUP BY object HAVING COUNT(metadata)=?)".format(seq=",".join(['?']*len(metadata)))
+                result = db.execute(sql, params).fetchall()
+                if return_metadata:
+                    names = {}
+                    for dbid, name, uri in result:
+                        metadata = {m[0] for m in db.execute("SELECT metadata FROM pyro_metadata WHERE object=?", (dbid,)).fetchall()}
+                        names[name] = uri, metadata
+                else:
+                    names = {name: uri for (dbid, name, uri) in result}
+                return names
+        except sqlite3.DatabaseError as e:
+            raise NamingError("sqlite error in optimized_metadata_search: "+str(e))
+
     def remove_items(self, items):
         try:
-            items = [(item,) for item in items]
             with closing(sqlite3.connect(self.dbfile)) as db:
-                db.executemany("DELETE FROM pyro_names WHERE name=?", items)
+                db.execute("PRAGMA foreign_keys=ON")
+                for item in items:
+                    dbid = db.execute("SELECT id FROM pyro_names WHERE name=?", (item,)).fetchone()
+                    if dbid:
+                        dbid = dbid[0]
+                        db.execute("DELETE FROM pyro_metadata WHERE object=?", (dbid,))
+                        db.execute("DELETE FROM pyro_names WHERE id=?", (dbid,))
                 db.commit()
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in remove_items: "+str(e))
 
-    def everything(self):
+    def everything(self, return_metadata=False):
         try:
             with closing(sqlite3.connect(self.dbfile)) as db:
-                result = db.execute("SELECT name, uri FROM pyro_names")
                 names = {}
-                for name, uri in result.fetchall():
-                    names[name] = uri
+                if return_metadata:
+                    for dbid, name, uri in db.execute("SELECT id, name, uri FROM pyro_names").fetchall():
+                        metadata = {m[0] for m in db.execute("SELECT metadata FROM pyro_metadata WHERE object=?", (dbid,)).fetchall()}
+                        names[name] = uri, metadata
+                else:
+                    for name, uri in db.execute("SELECT name, uri FROM pyro_names").fetchall():
+                        names[name] = uri
                 return names
         except sqlite3.DatabaseError as e:
             raise NamingError("sqlite error in everything: "+str(e))
@@ -160,6 +238,8 @@ class DbmStorage(MutableMapping):
     Dbm files cannot be accessed concurrently, so a strict concurrency model
     is used where only one operation is processed at the same time
     (this is very slow when compared to the in-memory storage)
+    DbmStorage does NOT support storing metadata! It only accepts empty metadata,
+    and always returns empty metadata.
     """
     def __init__(self, dbmfile):
         self.dbmfile = dbmfile
@@ -175,17 +255,20 @@ class DbmStorage(MutableMapping):
         with self.lock:
             try:
                 with closing(dbm.open(self.dbmfile)) as db:
-                    return db[item].decode("utf-8")
+                    return db[item].decode("utf-8"), frozenset()    # always return empty metadata
             except dbm.error as e:
                 raise NamingError("dbm error in getitem: "+str(e))
 
     def __setitem__(self, key, value):
+        uri, metadata = value
+        if metadata:
+            log.warning("DbmStorage doesn't support metadata, silently discarded")
         key = key.encode("utf-8")
-        value = value.encode("utf-8")
+        uri = uri.encode("utf-8")
         with self.lock:
             try:
                 with closing(dbm.open(self.dbmfile, "w")) as db:
-                    db[key] = value
+                    db[key] = uri
             except dbm.error as e:
                 raise NamingError("dbm error in setitem: "+str(e))
 
@@ -235,7 +318,7 @@ class DbmStorage(MutableMapping):
             except dbm.error as e:
                 raise NamingError("dbm error in clear: "+str(e))
 
-    def optimized_prefix_list(self, prefix):
+    def optimized_prefix_list(self, prefix, return_metadata=False):
         with self.lock:
             try:
                 with closing(dbm.open(self.dbmfile)) as db:
@@ -244,17 +327,19 @@ class DbmStorage(MutableMapping):
                         for key, value in db.items():
                             key = key.decode("utf-8")
                             if key.startswith(prefix):
-                                result[key] = value.decode("utf-8")
+                                uri = value.decode("utf-8")
+                                result[key] = (uri, frozenset()) if return_metadata else uri    # always return empty metadata
                     else:
                         for key in db.keys():
                             keystr = key.decode("utf-8")
                             if keystr.startswith(prefix):
-                                result[keystr] = db[key].decode("utf-8")
+                                uri = db[key].decode("utf-8")
+                                result[keystr] = (uri, frozenset()) if return_metadata else uri     # always return empty metadata
                     return result
             except dbm.error as e:
                 raise NamingError("dbm error in optimized_prefix_list: "+str(e))
 
-    def optimized_regex_list(self, regex):
+    def optimized_regex_list(self, regex, return_metadata=False):
         try:
             regex = re.compile(regex + "$")  # add end of string marker
         except re.error:
@@ -268,15 +353,22 @@ class DbmStorage(MutableMapping):
                         for key, value in db.items():
                             key = key.decode("utf-8")
                             if regex.match(key):
-                                result[key] = value.decode("utf-8")
+                                uri = value.decode("utf-8")
+                                result[key] = (uri, frozenset()) if return_metadata else uri     # always return empty metadata
                     else:
                         for key in db.keys():
                             keystr = key.decode("utf-8")
                             if regex.match(keystr):
-                                result[keystr] = db[key].decode("utf-8")
+                                uri = db[key].decode("utf-8")
+                                result[keystr] = (uri, frozenset()) if return_metadata else uri    # always return empty metadata
                     return result
             except dbm.error as e:
                 raise NamingError("dbm error in optimized_regex_list: "+str(e))
+
+    def optimized_metadata_search(self, metadata, return_metadata=False):
+        if metadata:
+            raise NamingError("DbmStorage doesn't support metadata")
+        return self.everything(return_metadata)
 
     def remove_items(self, items):
         with self.lock:
@@ -290,17 +382,19 @@ class DbmStorage(MutableMapping):
             except dbm.error as e:
                 raise NamingError("dbm error in remove_items: "+str(e))
 
-    def everything(self):
+    def everything(self, return_metadata=False):
         with self.lock:
             try:
                 with closing(dbm.open(self.dbmfile)) as db:
                     result = {}
                     if hasattr(db, "items"):
                         for key, value in db.items():
-                            result[key.decode("utf-8")] = value.decode("utf-8")
+                            uri = value.decode("utf-8")
+                            result[key.decode("utf-8")] = (uri, frozenset()) if return_metadata else uri    # always return empty metadata
                     else:
                         for key in db.keys():
-                            result[key.decode("utf-8")] = db[key].decode("utf-8")
+                            uri = db[key].decode("utf-8")
+                            result[key.decode("utf-8")] = (uri, frozenset()) if return_metadata else uri    # always return empty metadata
                     return result
             except dbm.error as e:
                 raise NamingError("dbm error in everything: "+str(e))